requirements.txt
obs-websocket-py
pyyaml